Change Room struct after database changed

+++ TeamCleaningOptimisation/CustomViews/TableViewCell.swift
@@ -16,7 +16,7 @@ class TableViewCell: UITableViewCell{
     @IBOutlet weak var roomIndexLabel: UILabel!
     @IBOutlet weak var timeLabel: UILabel!
     
-    func updateContent(roomID number: String, roomIndex index: Int) {
+    func updateContent(roomID number: String, roomIndex index: Double) {
         roomNumberLabel.text = number
         roomIndexLabel.text = String(index)
     }
